Move contents of core into __init__
To ensure all functions are accessible and visible from the top level,
simply move them into `__init__` and drop `core`. Tweak tests in light
of this change.
